update exercise magic numbers

+++ modules/30-variables/70-magic-numbers/main.cpp
@@ -5,8 +5,7 @@ int main() {
   int summ_of_temperatures { 588 };
   int numbers_of_hours { 24 };
 
-  float average_temperature { static_cast<float>(summ_of_temperatures)
-                              / numbers_of_hours };
+  int average_temperature { summ_of_temperatures / numbers_of_hours };
 
   std::cout << "Average daily temperature: "
             << average_temperature;
